agent/requirements.txt
strands-agents>=1.25.0
bedrock-agentcore>=1.2.0
